Can now only delete if you have the matching user_id to the recipe's one.

+++ app/Http/Controllers/RecipeController.php
@@ -85,8 +85,12 @@ class RecipeController extends Controller
         return redirect('/recipes')->with('success', 'Recipe added successfully!');
     }
     public function destroy(Recipe $recipe) {
-        $recipe->delete();
-        return redirect()->route('recipes.index')->with('success', 'Recipe deleted successfully!');
+        if ($recipe->user_id === auth()->user()->id)
+        {
+            $recipe->delete();
+            return redirect()->route('recipes.index')->with('success', 'Recipe deleted successfully!');
+        }
+        return back()->withError('Operation not permitted.');
     }
 
     public function edit($id) {
